requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
google-generativeai
anthropic
httpx